+++ tests/connections/test_unix_socket_connection.py
@@ -28,8 +28,10 @@ from gvm.connections import UnixSocketConnection, DEFAULT_TIMEOUT
 
 
 class DummyRequestHandler(socketserver.BaseRequestHandler):
+
     def handle(self):
-        response = bytes("<gmp_response status=\"200\" status_text=\"OK\"/>", 'utf-8')
+        response = bytes(
+            "<gmp_response status=\"200\" status_text=\"OK\"/>", 'utf-8')
         self.request.sendall(response)
 
 
@@ -40,10 +42,14 @@ class ThreadedUnixStreamServer(
 
 
 class UnixSocketConnectionTestCase(unittest.TestCase):
+
     def setUp(self):
-        self.socketname = "%s/%s.sock" % (tempfile.gettempdir(), str(uuid.uuid4()))
-        self.sockserv = ThreadedUnixStreamServer(self.socketname, DummyRequestHandler)
-        self.server_thread = threading.Thread(target=self.sockserv.serve_forever)
+        self.socketname = "%s/%s.sock" % (
+            tempfile.gettempdir(), str(uuid.uuid4()))
+        self.sockserv = ThreadedUnixStreamServer(
+            self.socketname, DummyRequestHandler)
+        self.server_thread = threading.Thread(
+            target=self.sockserv.serve_forever)
         self.server_thread.daemon = True
         self.server_thread.start()
 
@@ -53,20 +59,23 @@ class UnixSocketConnectionTestCase(unittest.TestCase):
         os.unlink(self.socketname)
 
     def test_unix_socket_connection_connect_read(self):
-        self.connection = UnixSocketConnection(self.socketname, DEFAULT_TIMEOUT)
+        self.connection = UnixSocketConnection(
+            path=self.socketname, timeout=DEFAULT_TIMEOUT)
         self.connection.connect()
         self.connection.read()
         self.connection.disconnect()
 
     def test_unix_socket_connection_connect_send_bytes_read(self):
-        self.connection = UnixSocketConnection(self.socketname, DEFAULT_TIMEOUT)
+        self.connection = UnixSocketConnection(
+            path=self.socketname, timeout=DEFAULT_TIMEOUT)
         self.connection.connect()
         self.connection.send(bytes("<gmp/>", 'utf-8'))
         self.connection.read()
         self.connection.disconnect()
 
     def test_unix_socket_connection_connect_send_str_read(self):
-        self.connection = UnixSocketConnection(self.socketname, DEFAULT_TIMEOUT)
+        self.connection = UnixSocketConnection(
+            path=self.socketname, timeout=DEFAULT_TIMEOUT)
         self.connection.connect()
         self.connection.send("<gmp/>")
         self.connection.read()