+++ CHANGELOG.md
@@ -1,17 +1,24 @@
 # python-gvm 1.0.0 (unreleased)
 
-## gvm.protocols.gmpv1
+## gvm.protocols.gmpv7
 
 * Fixed generating XML for help command
 * **help** method **type** argument got renamed to **help_type**
 * **help** method **help_type** argument will be checked for invalid values
 * Fixed wrong order of key and value for condition_data, event_data and
   method_data dict parameters of **create_alert** method.
+* **create_credential** requires a credential_type argument now.
+* Optional parameters are required to be passed as keyword arguments.
 
 ## gvm.xml
 
 * Added helper function to validate xml input **gvm.xml.validate_xml_string**
 * **pretty_print** accepts a xml string as input too
+
+## gvm.connections
+
+* Optional arguments for connection class constructors must be passed as
+  keyword arguments.
 
 # python-gvm 1.0.0.beta1 (13.11.2018)